requirements.txt
absl-py
einops
gin-config
GPUtil
librosa
numpy
pytorch_lightning
PyYAML
scikit_learn
scipy
torch
tqdm
udls
cached-conv
nn-tilde
torchaudio
tensorboard
pytest
Flask

